Relocate/capitalize `model.js` to live with associated routes
